Bumped TWO_TO_THREE_TRANSITION_COUNT due to ABI changes.

+++ cram/cram_samtools.h
@@ -80,13 +80,13 @@ int bam_construct_seq(bam_seq_t **bp, size_t extra_len,
                       const char *qname, size_t qname_len,
                       int flag,
                       int rname,      // Ref ID
-                      int pos,
-                      int end,        // aligned start/end coords
+                      int64_t pos,
+                      int64_t end,        // aligned start/end coords
                       int mapq,
                       uint32_t ncigar, const uint32_t *cigar,
                       int mrnm,       // Mate Ref ID
-                      int mpos,
-                      int isize,
+                      int64_t mpos,
+                      int64_t isize,
                       int len,
                       const char *seq,
                       const char *qual);
